feat: 게임/게임요청 권한 기능(권리자만 CRUD 가능) 추가
1.게임/게임요청 권한 기능(권리자만 CRUD 가능) 추가

+++ src/main/java/com/example/gamemate/domain/game/service/GameService.java
@@ -10,6 +10,8 @@ import com.example.gamemate.domain.review.dto.ReviewFindByAllResponseDto;
 import com.example.gamemate.domain.review.entity.Review;
 import com.example.gamemate.domain.review.repository.ReviewRepository;
 import com.example.gamemate.domain.user.entity.User;
+import com.example.gamemate.domain.user.enums.Role;
+import com.example.gamemate.global.constant.ErrorCode;
 import com.example.gamemate.global.exception.ApiException;
 import com.example.gamemate.global.s3.S3Service;
 import lombok.RequiredArgsConstructor;
@@ -26,8 +28,6 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
 import java.util.List;
 
-import static com.example.gamemate.global.constant.ErrorCode.GAME_NOT_FOUND;
-
 
 @Service
 @Slf4j
@@ -39,8 +39,12 @@ public class GameService {
     private final GameImageRepository gameImageRepository;
 
     @Transactional
-    public GameCreateResponseDto createGame(GameCreateRequestDto gameCreateRequestDto , MultipartFile file) {
+    public GameCreateResponseDto createGame(User loginUser, GameCreateRequestDto gameCreateRequestDto, MultipartFile file) {
 
+        //관리자만 가능함(생성)
+        if (!loginUser.getRole().equals(Role.ADMIN)) {
+            throw new ApiException(ErrorCode.FORBIDDEN);
+        }
         // 게임 엔티티 생성
         Game game = new Game(
                 gameCreateRequestDto.getTitle(),
@@ -75,43 +79,39 @@ public class GameService {
     }
 
     @Transactional
-    public GameFindByIdResponseDto findGameById(User longinUser, Long id) {
-
-        String nickName = longinUser.getNickname();
+    public GameFindByIdResponseDto findGameById(User loginUser, Long id) {
 
         Game game = gameRepository.findGameById(id)
-                .orElseThrow(() -> new ApiException(GAME_NOT_FOUND));
+                .orElseThrow(() -> new ApiException(ErrorCode.GAME_NOT_FOUND));
 
-        Pageable pageable = PageRequest.of(0, 5, Sort.by("createdAt").descending());
-        Page<Review> reviewPage = reviewRepository.findAllByGame(game, pageable);
-
-        // Review를 ReviewFindByAllResponseDto로 변환하면서 닉네임 추가
-        Page<ReviewFindByAllResponseDto> reviews = reviewPage.map(review ->
-                new ReviewFindByAllResponseDto(review, longinUser.getNickname())
-        );
-
-        return new GameFindByIdResponseDto(game, reviews, nickName);
+        return new GameFindByIdResponseDto(game);
     }
 
     @Transactional
-    public void updateGame(Long id, GameUpdateRequestDto requestDto, MultipartFile newFile) {
+    public void updateGame(Long id, GameUpdateRequestDto requestDto, MultipartFile newFile, User loginUser) {
+
+        //관리자만 가능함(수정)
+        if (!loginUser.getRole().equals(Role.ADMIN)) {
+            throw new ApiException(ErrorCode.FORBIDDEN);
+        }
+
         Game game = gameRepository.findGameById(id)
-                .orElseThrow(() -> new ApiException(GAME_NOT_FOUND));
+                .orElseThrow(() -> new ApiException(ErrorCode.GAME_NOT_FOUND));
 
         // 기존 파일이 있고 새 파일이 업로드된 경우
-            // 1. 기존 S3 파일 삭제
-//        if (!game.getImages().isEmpty()) {
-//            for (GameImage image : game.getImages()) {
-//                s3Service.deleteFile(image.getFilePath());
-//            }
-//        }
+        // 1. 기존 S3 파일 삭제
+        if (!game.getImages().isEmpty()) {
+            for (GameImage image : game.getImages()) {
+                s3Service.deleteFile(image.getFilePath());
+            }
+        }
 
         List<GameImage> gameImages = gameImageRepository.findGameImagesByGameId(id);
         if (!gameImages.isEmpty()) {
             gameImageRepository.deleteAll(gameImages);
         }
 
-            // 2. 새 파일 업로드
+        // 2. 새 파일 업로드
         if (newFile != null && !newFile.isEmpty()) {
             try {
                 String fileUrl = s3Service.uploadFile(newFile);
@@ -137,9 +137,15 @@ public class GameService {
     }
 
     @Transactional
-    public void deleteGame(Long id) {
+    public void deleteGame(Long id, User loginUser) {
+
+        //관리자만 가능함(삭제)
+        if (!loginUser.getRole().equals(Role.ADMIN)) {
+            throw new ApiException(ErrorCode.FORBIDDEN);
+        }
+
         Game game = gameRepository.findGameById(id)
-                .orElseThrow(() -> new ApiException(GAME_NOT_FOUND));
+                .orElseThrow(() -> new ApiException(ErrorCode.GAME_NOT_FOUND));
 
         // 게임에 연결된 모든 이미지 삭제
         if (!game.getImages().isEmpty()) {
